requirements.txt
# FastAPI and ASGI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
starlette==0.35.1

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
greenlet==3.0.3

# Redis and Celery
celery[redis]==5.3.6
redis==5.0.1
flower==2.0.1

# Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
cryptography==41.0.7
bcrypt==4.1.2

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Validation and Settings
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Templating
jinja2==3.1.3
itsdangerous==2.1.2

# Utilities
python-dotenv==1.0.0
pytz==2024.1
python-dateutil==2.8.2
structlog==24.1.0

# OAuth
authlib==1.3.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
